+++ R/extended.shared.R
@@ -50,7 +50,8 @@ extended.shared.pedigreeList <- function (id, rho=1, theta=1, ...)
     for (i in 1:length(famlist)) {
       family <- id[i]
       temp <- extended.shared(family)
-      matlist[[i]] <- as(Matrix::forceSymmetric(temp), "dsCMatrix")
+     #  matlist[[i]] <- as(Matrix::forceSymmetric(temp), "dsCMatrix")
+      matlist[[i]] <- as(as(as(Matrix::forceSymmetric(temp), "dMatrix"), "symmetricMatrix"), "CsparseMatrix")            
     }
     result <- Matrix::bdiag(matlist)