+++ R/founder.env.R
@@ -47,7 +47,8 @@ founder.shared.pedigreeList <- function (id, ...)
     for (i in 1:length(famlist)) {
       family <- id[i]
       temp <- founder.shared(family)
-      matlist[[i]] <- as(Matrix::forceSymmetric(temp), "dsCMatrix")
+      # matlist[[i]] <- as(Matrix::forceSymmetric(temp), "dsCMatrix")
+      matlist[[i]] <- as(as(as(Matrix::forceSymmetric(temp), "dMatrix"), "symmetricMatrix"), "CsparseMatrix")            
     }
     result <- Matrix::bdiag(matlist)